Add export history to export
